Move test utils to src and add RPC script
